+++ backend/src/modules/document-management/models/TramiteEliminacion.model.js
@@ -1,5 +1,4 @@
-import { DataTypes } from "sequelize";
-import { sequelize } from "../config/db.config.js";
+import { sequelize } from "../../../config/db.config.js";
 
 export const TramiteEliminacion = sequelize.define(
   "tramite_eliminacion",
